python-sidecar/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
httpx>=0.25.0
sounddevice>=0.4.6
numpy>=1.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
openai>=1.0.0
elevenlabs>=0.2.0
deepgram-sdk>=3.0.0
